+++ tests/test_aev.py
@@ -153,20 +153,6 @@ class TestAEV(_TestAEVBase):
             start += conformations
             self.assertAEVEqual(expected_radial, expected_angular, aev_)
 
-    def testNIST(self):
-        datafile = os.path.join(path, 'test_data/NIST/all')
-        with open(datafile, 'rb') as f:
-            data = pickle.load(f)
-            for coordinates, species, radial, angular, _, _ in data:
-                if self.random_skip():
-                    continue
-                coordinates = torch.from_numpy(coordinates).to(torch.float)
-                species = torch.from_numpy(species)
-                radial = torch.from_numpy(radial).to(torch.float)
-                angular = torch.from_numpy(angular).to(torch.float)
-                _, aev = self.aev_computer((species, coordinates))
-                self.assertAEVEqual(radial, angular, aev)
-
     @unittest.skipIf(not torch.cuda.is_available(), "Too slow on CPU")
     def testGradient(self):
         """Test validity of autodiff by comparing analytical and numerical
