+++ tests/test_energies.py
@@ -42,26 +42,6 @@ class TestEnergies(unittest.TestCase):
                 max_diff = (energies - energies_).abs().max().item()
                 self.assertLess(max_diff, self.tolerance)
 
-    def testBenzeneMD(self):
-        tolerance = 1e-5
-        for i in range(10):
-            datafile = os.path.join(path, 'test_data/benzene-md/{}.dat'.format(i))
-            with open(datafile, 'rb') as f:
-                coordinates, species, _, _, energies, _, cell, pbc \
-                    = pickle.load(f)
-                coordinates = torch.from_numpy(coordinates).float().unsqueeze(0)
-                species = torch.from_numpy(species).unsqueeze(0)
-                cell = torch.from_numpy(cell).float()
-                pbc = torch.from_numpy(pbc)
-                coordinates = torchani.utils.map2central(cell, coordinates, pbc)
-                coordinates = self.transform(coordinates)
-                species = self.transform(species)
-                energies = self.transform(energies)
-                _, aev = self.aev_computer((species, coordinates), cell=cell, pbc=pbc)
-                _, energies_ = self.nn((species, aev))
-                max_diff = (energies - energies_).abs().max().item()
-                self.assertLess(max_diff, tolerance)
-
     def testTripeptideMD(self):
         tolerance = 2e-4
         for i in range(100):
@@ -99,21 +79,6 @@ class TestEnergies(unittest.TestCase):
         _, energies_ = self.model((species_coordinates['species'], species_coordinates['coordinates']))
         max_diff = (energies - energies_).abs().max().item()
         self.assertLess(max_diff, self.tolerance)
-
-    def testNIST(self):
-        datafile = os.path.join(path, 'test_data/NIST/all')
-        with open(datafile, 'rb') as f:
-            data = pickle.load(f)
-            for coordinates, species, _, _, e, _ in data:
-                if self.random_skip():
-                    continue
-                coordinates = torch.from_numpy(coordinates).to(torch.float)
-                species = torch.from_numpy(species)
-                energies = torch.from_numpy(e).to(torch.float)
-                _, energies_ = self.model((species, coordinates))
-                natoms = coordinates.shape[1]
-                max_diff = (energies - energies_).abs().max().item()
-                self.assertLess(max_diff / math.sqrt(natoms), self.tolerance)
 
 
 class TestEnergiesEnergyShifterJIT(TestEnergies):
